pyproject.toml
[tool.poetry]
name = "faiss"
version = "0.1.0"
description = ""
authors = ["yano0 <yanonay3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
faiss-cpu = "^1.7.2"
torch = {url = "https://download.pytorch.org/whl/cu113/torch-1.12.1%2Bcu113-cp38-cp38-linux_x86_64.whl"}
transformers = "^4.23.1"
numpy = "^1.23.4"
pandas = "^1.5.1"
fugashi = "^1.2.0"
unidic = "^1.1.0"
ipadic = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.16.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
